Impelement noGitImplementation to use frolint without git integration
